feat: improve auth service

+++ app/src/services/base-api-service.ts
@@ -1,3 +1,5 @@
+import {AuthService} from './auth-service';
+
 export interface RequestOptions {
     abort?: AbortSignal;
     headers?: Record<string, string>;
@@ -5,8 +7,13 @@ export interface RequestOptions {
 }
 
 export class BaseApiService {
+    private readonly auth = new AuthService();
+
     public async get<T>(path: string, options?: RequestOptions): Promise<T> {
-        const accessToken = await fetchApiToken(options?.abort);
+        const accessToken = await this.auth.getAccessToken(options?.abort);
+        if (accessToken == null) {
+            throw new Error('No valid access token available');
+        }
 
         const response = await fetch(combineUrl(path, options), {
             method: 'GET',
@@ -16,7 +23,7 @@ export class BaseApiService {
 
         if (response.status !== 200) {
             if (response.status === 401) {
-                resetLocalStorageJwt();
+                this.auth.logout();
             }
             throw new Error(`GET ${path} failed with status ${response.status}`);
         }
@@ -32,9 +39,6 @@ export class BaseApiService {
         });
 
         if (response.status !== 200) {
-            if (response.status === 401) {
-                resetLocalStorageJwt();
-            }
             throw new Error(`GET ${path} failed with status ${response.status}`);
         }
 
@@ -42,7 +46,10 @@ export class BaseApiService {
     }
 
     public async getBlob(path: string, options?: RequestOptions): Promise<Blob> {
-        const accessToken = await fetchApiToken(options?.abort);
+        const accessToken = await this.auth.getAccessToken(options?.abort);
+        if (accessToken == null) {
+            throw new Error('No valid access token available');
+        }
 
         const response = await fetch(combineUrl(path, options), {
             method: 'GET',
@@ -52,7 +59,7 @@ export class BaseApiService {
 
         if (response.status !== 200) {
             if (response.status === 401) {
-                resetLocalStorageJwt();
+                this.auth.logout();
             }
             throw new Error(`GET ${path} failed with status ${response.status}`);
         }
@@ -68,9 +75,6 @@ export class BaseApiService {
         });
 
         if (response.status !== 200) {
-            if (response.status === 401) {
-                resetLocalStorageJwt();
-            }
             throw new Error(`GET ${path} failed with status ${response.status}`);
         }
 
@@ -78,7 +82,10 @@ export class BaseApiService {
     }
 
     public async post<T, R>(path: string, body: T, options?: RequestOptions): Promise<R> {
-        const accessToken = await fetchApiToken();
+        const accessToken = await this.auth.getAccessToken(options?.abort);
+        if (accessToken == null) {
+            throw new Error('No valid access token available');
+        }
 
         const response = await fetch(combineUrl(path, options), {
             method: 'POST',
@@ -89,7 +96,7 @@ export class BaseApiService {
 
         if (response.status !== 200 && response.status !== 201) {
             if (response.status === 401) {
-                resetLocalStorageJwt();
+                this.auth.logout();
             }
             throw new Error(`POST ${path} failed with status ${response.status}`);
         }
@@ -106,9 +113,6 @@ export class BaseApiService {
         });
 
         if (response.status !== 200 && response.status !== 201) {
-            if (response.status === 401) {
-                resetLocalStorageJwt();
-            }
             throw new Error(`POST ${path} failed with status ${response.status}`);
         }
 
@@ -116,7 +120,10 @@ export class BaseApiService {
     }
 
     public async postFormData<R>(path: string, formData: FormData, options?: RequestOptions): Promise<R> {
-        const accessToken = await fetchApiToken(options?.abort);
+        const accessToken = await this.auth.getAccessToken(options?.abort);
+        if (accessToken == null) {
+            throw new Error('No valid access token available');
+        }
 
         const response = await fetch(combineUrl(path, options), {
             method: 'POST',
@@ -127,7 +134,7 @@ export class BaseApiService {
 
         if (response.status !== 200 && response.status !== 201) {
             if (response.status === 401) {
-                resetLocalStorageJwt();
+                this.auth.logout();
             }
             throw new Error(`POST ${path} failed with status ${response.status}`);
         }
@@ -144,9 +151,6 @@ export class BaseApiService {
         });
 
         if (response.status !== 200 && response.status !== 201) {
-            if (response.status === 401) {
-                resetLocalStorageJwt();
-            }
             throw new Error(`POST ${path} failed with status ${response.status}`);
         }
 
@@ -154,7 +158,10 @@ export class BaseApiService {
     }
 
     public async postFormUrlEncoded<R>(path: string, formData: URLSearchParams, options?: RequestOptions): Promise<R> {
-        const accessToken = await fetchApiToken(options?.abort);
+        const accessToken = await this.auth.getAccessToken(options?.abort);
+        if (accessToken == null) {
+            throw new Error('No valid access token available');
+        }
 
         const response = await fetch(combineUrl(path, options), {
             method: 'POST',
@@ -168,7 +175,7 @@ export class BaseApiService {
 
         if (response.status !== 200 && response.status !== 201) {
             if (response.status === 401) {
-                resetLocalStorageJwt();
+                this.auth.logout();
             }
             throw new Error(`POST ${path} failed with status ${response.status}`);
         }
@@ -187,9 +194,6 @@ export class BaseApiService {
         });
 
         if (response.status !== 200 && response.status !== 201) {
-            if (response.status === 401) {
-                resetLocalStorageJwt();
-            }
             throw new Error(`POST ${path} failed with status ${response.status}`);
         }
 
@@ -197,7 +201,10 @@ export class BaseApiService {
     }
 
     public async put<T, R>(path: string, body: T, options?: RequestOptions): Promise<R> {
-        const accessToken = await fetchApiToken();
+        const accessToken = await this.auth.getAccessToken(options?.abort);
+        if (accessToken == null) {
+            throw new Error('No valid access token available');
+        }
 
         const response = await fetch(combineUrl(path, options), {
             method: 'PUT',
@@ -208,7 +215,7 @@ export class BaseApiService {
 
         if (response.status !== 200) {
             if (response.status === 401) {
-                resetLocalStorageJwt();
+                this.auth.logout();
             }
             throw new Error(`PUT ${path} failed with status ${response.status}`);
         }
@@ -217,7 +224,10 @@ export class BaseApiService {
     }
 
     public async delete(path: string, options?: RequestOptions): Promise<void> {
-        const accessToken = await fetchApiToken(options?.abort);
+        const accessToken = await this.auth.getAccessToken(options?.abort);
+        if (accessToken == null) {
+            throw new Error('No valid access token available');
+        }
 
         const response = await fetch(combineUrl(path, options), {
             method: 'DELETE',
@@ -227,115 +237,11 @@ export class BaseApiService {
 
         if (response.status !== 200 && response.status !== 204) {
             if (response.status === 401) {
-                resetLocalStorageJwt();
+                this.auth.logout();
             }
             throw new Error(`DELETE ${path} failed with status ${response.status}`);
         }
     }
-}
-
-export const STORAGE_KEY_JWT = 'api-jwt';
-const EXPIRATION_PADDING_MS = 30;
-
-interface JWT {
-    access: {
-        token: string;
-        expires: number; // Unix timestamp in seconds
-    };
-    refresh: {
-        token: string;
-        expires: number; // Unix timestamp in seconds
-    };
-}
-
-async function fetchApiToken(signal?: AbortSignal): Promise<string> {
-    let jwt = getLocalStorageJwt();
-
-    // If there is no JWT in local storage, or if the refresh token has expired, clear local storage and throw an error
-    if (jwt == null) {
-        resetLocalStorageJwt();
-        throw new Error();
-    }
-
-    if (jwt.access.expires <= Date.now()) {
-        jwt = await getApiJWT(jwt.refresh.token, signal);
-    }
-
-    if (jwt == null) {
-        resetLocalStorageJwt();
-        throw new Error();
-    }
-
-    return jwt.access.token;
-}
-
-export function getLocalStorageJwt(): JWT | null {
-    const jwtStr = localStorage.getItem(STORAGE_KEY_JWT);
-    if (jwtStr == null) {
-        return null;
-    }
-
-    let jwt: JWT | null = null;
-    try {
-        jwt = JSON.parse(jwtStr) as JWT;
-    } catch {
-        return null;
-    }
-
-    if (jwt.refresh.expires <= Date.now()) {
-        return null;
-    }
-
-    return jwt;
-}
-
-async function getApiJWT(refreshToken: string, signal?: AbortSignal): Promise<JWT | null> {
-    const url = `${AppConfig.oidc.hostname}/realms/${AppConfig.oidc.realm}/protocol/openid-connect/token`;
-    const body = new URLSearchParams({
-        grant_type: 'refresh_token',
-        client_id: AppConfig.oidc.client,
-        refresh_token: refreshToken,
-    });
-    const headers = {
-        'Content-Type': 'application/x-www-form-urlencoded;charset=UTF-8',
-    };
-
-    const response = await window.fetch(url, {
-        method: 'POST',
-        body: body,
-        headers: headers,
-        signal: signal,
-    });
-
-    if (response.status !== 200) {
-        return null;
-    }
-
-    const data = await response
-        .json();
-
-    const jwt: JWT = {
-        access: {
-            token: data.access_token,
-            expires: (Date.now() + (data.expires_in * 1000)) - EXPIRATION_PADDING_MS,
-        },
-        refresh: {
-            token: data.refresh_token,
-            expires: (Date.now() + (data.refresh_expires_in * 1000)) - EXPIRATION_PADDING_MS,
-        },
-    };
-
-    storeLocalStorageJwt(jwt);
-
-    return jwt;
-}
-
-export function storeLocalStorageJwt(jwt: JWT) {
-    localStorage.setItem(STORAGE_KEY_JWT, JSON.stringify(jwt));
-}
-
-export function resetLocalStorageJwt() {
-    localStorage.removeItem(STORAGE_KEY_JWT);
 }
 
 function combineUrl(path: string, options?: RequestOptions): string {
